requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.
anyjson>=0.3.3
httplib2>=0.7.5
lxml>=2.3
oslo.i18n>=1.5.0,<1.6.0  # Apache-2.0
oslo.utils>=1.4.0,<1.5.0                       # Apache-2.0
pbr>=0.6,!=0.7,<1.0
PrettyTable>=0.7,<0.8
python-keystoneclient>=1.1.0
python-openstackclient>=1.0.0
six>=1.9.0
